setcookie('key', 'value', time() );
Cookies are a mechanism for storing data in the remote browser and this tracking or identifying return users. You can set specific data to be stored in the browserr, and then retrieve it when the user visits the site again. 

It does not have to store valuable information. for that we have sessions